Implemented a few more things, creating UI in the app

+++ network/src/main/java/com/bogtech/network/feed/model/subtypes/Amount.kt
@@ -2,5 +2,5 @@ package com.bogtech.network.feed.model.subtypes
 
 data class Amount (
     val currency: String,
-    val minorUnits: Int,
+    val minorUnits: Double,
 )
